211214_01_글 수정, 삭제 기능 추가

+++ src/main/webapp/view.jsp
@@ -14,7 +14,7 @@
 <body>
 	<%
 	String userID = "tlstmdgjs";
-	if(session.getAttribute("userID") != null) {
+	if (session.getAttribute("userID") != null) {
 		userID = (String) session.getAttribute("userID");
 	}
 	int bbsID = 0;
@@ -62,7 +62,8 @@
 				<tbody>
 					<tr>
 						<td style="width: 20%;">글 제목</td>
-						<td colspan="2"><%=bbs.getBbsTitle().replaceAll(" ", "&nbsp;").replaceAll("<", "&lt;").replaceAll(">", "&gt;").replaceAll("\n", "<br>")%></td>
+						<td colspan="2"><%=bbs.getBbsTitle().replaceAll(" ", "&nbsp;").replaceAll("<", "&lt;").replaceAll(">", "&gt;").replaceAll("\n",
+		"<br>")%></td>
 					</tr>
 					<tr>
 						<td>작성자</td>
@@ -71,22 +72,23 @@
 					<tr>
 						<td>작성일자</td>
 						<td colspan="2"><%=bbs.getBbsDate().substring(0, 11) + bbs.getBbsDate().substring(11, 13) + "시"
-		+ bbs.getBbsDate().substring(14, 16) + "분" %></td>
+		+ bbs.getBbsDate().substring(14, 16) + "분"%></td>
 					</tr>
 					<tr>
 						<td>내용</td>
-						<td colspan="2" style="height: 200px text-align: left;"><%= bbs.getBbsContent().replaceAll(" ", "&nbsp;").replaceAll("<", "&lt;").replaceAll(">", "&gt;").replaceAll("\n", "<br>") %></td>
+						<td colspan="2" style="height: 200px text-align: left;"><%=bbs.getBbsContent().replaceAll(" ", "&nbsp;").replaceAll("<", "&lt;").replaceAll(">", "&gt;").replaceAll("\n",
+		"<br>")%></td>
 					</tr>
 				</tbody>
 			</table>
 			<a href="bs.jsp" class="btn btn-primary">목록</a>
 			<%
-				if(userID != null && userID.equals(bbs.getUserID())) {
+			if (userID != null && userID.equals(bbs.getUserID())) {
 			%>
-					<a href="update.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">수정</a>
-					<a href="deleteAction.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">삭제</a>
+			<a href="update.jsp?bbsID=<%=bbsID%>" class="btn btn-primary">수정</a>
+			<a onclick="return confirm('정말로 삭제하시겠습니까?')" href="deleteAction.jsp?bbsID=<%=bbsID%>" class="btn btn-primary">삭제</a>
 			<%
-				}
+			}
 			%>
 		</div>
 	</div>
